Feature/EC-93 ss   income statements running bal

Approved-by: Liana Wiehahn
Approved-by: Ruald de Bruyn

+++ src/api/Libraries/Modules/ECDLink.DataAccessLayer/Entities/IncomeStatements/StatementsIncomeStatement.cs
@@ -20,9 +20,9 @@ namespace ECDLink.DataAccessLayer.Entities.IncomeStatements
         public int Year { get; set; }
         public int Period { get; set; }
         public bool Submitted { get; set; }
-        public decimal IncomeTotal { get; set; }
-        public decimal ExpenseTotal { get; set; }
-        public decimal Balance { get; set; }
+        public double IncomeTotal { get; set; }
+        public double ExpenseTotal { get; set; }
+        public double Balance { get; set; }
         public DateTime? SubmittedDate { get; set; }
         public string UserId { get; set; }
